FIX: Correct order in tests writer
