+++ tests/test_query_index_align.cc
@@ -1,37 +1,14 @@
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/catch_approx.hpp>
 
-#include <algorithm>
 #include <sstream>
 #include <vector>
 
-#include "output_writer.h"
 #include "parameters.h"
 #include "query_index_align.h"
 
+#include "dummy_output_writer.h"
 #include "testing_helpers.h"
-
-class DummyOutputWriter: public BaseOutputWriter {
-public:
-    void display(const AlignStats &r, std::ostream& output) const override {
-        m_stats.push_back(r);
-    }
-    void display(const std::vector<AlignStats>& results, std::ostream& output) const override {
-        m_stats.resize(m_stats.size() + results.size());
-        std::copy(results.begin(), results.end(), m_stats.rbegin());
-    }
-
-    size_t ncalls() const {
-        return m_stats.size();
-    }
-
-    AlignStats operator[](size_t i) const {
-        return m_stats[i];
-    }
-
-private:
-    mutable std::vector<AlignStats> m_stats;
-};
 
 
 TEST_CASE("Query the index in align mode", "[query-align]") {
@@ -51,10 +28,10 @@ TEST_CASE("Query the index in align mode", "[query-align]") {
 
         // Then
         REQUIRE(writer.ncalls() == 2);
-        REQUIRE(writer[0].subjectId == "A");
-        REQUIRE(writer[0].pctIdentity == Catch::Approx(57.14).margin(0.01));
-        REQUIRE(writer[1].subjectId == "B");
-        REQUIRE(writer[1].pctIdentity == Catch::Approx(83.33).margin(0.01));
+        REQUIRE(writer[0].subjectId == "B");
+        REQUIRE(writer[0].pctIdentity == Catch::Approx(83.33).margin(0.01));
+        REQUIRE(writer[1].subjectId == "A");
+        REQUIRE(writer[1].pctIdentity == Catch::Approx(57.14).margin(0.01));
     }
 
     SECTION("vptree") {
@@ -72,10 +49,10 @@ TEST_CASE("Query the index in align mode", "[query-align]") {
 
         // Then
         REQUIRE(writer.ncalls() == 2);
-        REQUIRE(writer[0].subjectId == "A");
-        REQUIRE(writer[0].pctIdentity == Catch::Approx(57.14).margin(0.01));
-        REQUIRE(writer[1].subjectId == "B");
-        REQUIRE(writer[1].pctIdentity == Catch::Approx(83.33).margin(0.01));
+        REQUIRE(writer[0].subjectId == "B");
+        REQUIRE(writer[0].pctIdentity == Catch::Approx(83.33).margin(0.01));
+        REQUIRE(writer[1].subjectId == "A");
+        REQUIRE(writer[1].pctIdentity == Catch::Approx(57.14).margin(0.01));
     }
 
     SECTION("hnsw") {
@@ -93,9 +70,9 @@ TEST_CASE("Query the index in align mode", "[query-align]") {
 
         // Then
         REQUIRE(writer.ncalls() == 2);
-        REQUIRE(writer[0].subjectId == "A");
-        REQUIRE(writer[0].pctIdentity == Catch::Approx(57.14).margin(0.01));
-        REQUIRE(writer[1].subjectId == "B");
-        REQUIRE(writer[1].pctIdentity == Catch::Approx(83.33).margin(0.01));
+        REQUIRE(writer[0].subjectId == "B");
+        REQUIRE(writer[0].pctIdentity == Catch::Approx(83.33).margin(0.01));
+        REQUIRE(writer[1].subjectId == "A");
+        REQUIRE(writer[1].pctIdentity == Catch::Approx(57.14).margin(0.01));
     }
 }
